Move all built-in matchers into the default matcher extension
