+++ src/PhpSpec/Extensions/DefaultMatchers/Matcher/ObjectStateMatcher.php
@@ -11,11 +11,12 @@
  * file that was distributed with this source code.
  */
 
-namespace PhpSpec\Matcher;
+namespace PhpSpec\Extensions\DefaultMatchers\Matcher;
 
 use PhpSpec\Formatter\Presenter\Presenter;
 use PhpSpec\Exception\Example\FailureException;
 use PhpSpec\Exception\Fracture\MethodNotFoundException;
+use PhpSpec\Matcher\Matcher;
 
 final class ObjectStateMatcher implements Matcher
 {
