+++ src/PhpSpec/Extensions/DefaultMatchers/Matcher/StringContainMatcher.php
@@ -11,12 +11,12 @@
  * file that was distributed with this source code.
  */
 
-namespace PhpSpec\Matcher;
+namespace PhpSpec\Extensions\DefaultMatchers\Matcher;
 
-use PhpSpec\Formatter\Presenter\Presenter;
 use PhpSpec\Exception\Example\FailureException;
+use PhpSpec\Formatter\Presenter\Presenter;
 
-final class StringEndMatcher extends BasicMatcher
+final class StringContainMatcher extends BasicMatcher
 {
     /**
      * @var Presenter
@@ -32,58 +32,43 @@ final class StringEndMatcher extends BasicMatcher
     }
 
     /**
-     * @param string $name
-     * @param mixed  $subject
-     * @param array  $arguments
-     *
-     * @return bool
+     * {@inheritdoc}
      */
-    public function supports(string $name, $subject, array $arguments): bool
+    public function supports(string $name, $subject, array $arguments) : bool
     {
-        return 'endWith' === $name
+        return 'contain' === $name
             && \is_string($subject)
-            && 1 == \count($arguments)
-        ;
+            && 1 === \count($arguments)
+            && \is_string($arguments[0]);
     }
 
     /**
-     * @param mixed $subject
-     * @param array $arguments
-     *
-     * @return bool
+     * {@inheritdoc}
      */
-    protected function matches($subject, array $arguments): bool
+    protected function matches($subject, array $arguments) : bool
     {
-        return $arguments[0] === substr($subject, 0 - \strlen($arguments[0]));
+        return false !== strpos($subject, $arguments[0]);
     }
 
     /**
-     * @param string $name
-     * @param mixed  $subject
-     * @param array  $arguments
-     *
-     * @return FailureException
+     * {@inheritdoc}
      */
     protected function getFailureException(string $name, $subject, array $arguments): FailureException
     {
         return new FailureException(sprintf(
-            'Expected %s to end with %s, but it does not.',
+            'Expected %s to contain %s, but it does not.',
             $this->presenter->presentString($subject),
             $this->presenter->presentString($arguments[0])
         ));
     }
 
     /**
-     * @param string $name
-     * @param mixed  $subject
-     * @param array  $arguments
-     *
-     * @return FailureException
+     * {@inheritdoc}
      */
     protected function getNegativeFailureException(string $name, $subject, array $arguments): FailureException
     {
         return new FailureException(sprintf(
-            'Expected %s not to end with %s, but it does.',
+            'Expected %s not to contain %s, but it does.',
             $this->presenter->presentString($subject),
             $this->presenter->presentString($arguments[0])
         ));
