+++ src/PhpSpec/Extensions/DefaultMatchers/Matcher/StringStartMatcher.php
@@ -11,13 +11,12 @@
  * file that was distributed with this source code.
  */
 
-namespace PhpSpec\Matcher;
+namespace PhpSpec\Extensions\DefaultMatchers\Matcher;
 
 use PhpSpec\Formatter\Presenter\Presenter;
 use PhpSpec\Exception\Example\FailureException;
-use ArrayAccess;
 
-final class TraversableKeyMatcher extends BasicMatcher
+final class StringStartMatcher extends BasicMatcher
 {
     /**
      * @var Presenter
@@ -33,51 +32,60 @@ final class TraversableKeyMatcher extends BasicMatcher
     }
 
     /**
-     * {@inheritdoc}
+     * @param string $name
+     * @param mixed  $subject
+     * @param array  $arguments
+     *
+     * @return bool
      */
     public function supports(string $name, $subject, array $arguments): bool
     {
-        return 'haveKey' === $name
-            && 1 === \count($arguments)
-            && $subject instanceof \Traversable
+        return 'startWith' === $name
+            && \is_string($subject)
+            && 1 == \count($arguments)
         ;
     }
 
     /**
-     * {@inheritdoc}
+     * @param mixed $subject
+     * @param array $arguments
+     *
+     * @return bool
      */
     protected function matches($subject, array $arguments): bool
     {
-        foreach ($subject as $key => $value) {
-            if ($key === $arguments[0]) {
-                return true;
-            }
-        }
-
-        return false;
+        return 0 === strpos($subject, $arguments[0]);
     }
 
     /**
-     * {@inheritdoc}
+     * @param string $name
+     * @param mixed  $subject
+     * @param array  $arguments
+     *
+     * @return FailureException
      */
     protected function getFailureException(string $name, $subject, array $arguments): FailureException
     {
         return new FailureException(sprintf(
-            'Expected %s to have %s key, but it does not.',
-            $this->presenter->presentValue($subject),
-            $this->presenter->presentValue($arguments[0])
+            'Expected %s to start with %s, but it does not.',
+            $this->presenter->presentString($subject),
+            $this->presenter->presentString($arguments[0])
         ));
     }
 
     /**
-     * {@inheritdoc}
+     * @param string $name
+     * @param mixed  $subject
+     * @param array  $arguments
+     *
+     * @return FailureException
      */
     protected function getNegativeFailureException(string $name, $subject, array $arguments): FailureException
     {
         return new FailureException(sprintf(
-            'Expected %s not to have %s key, but it does.',
-            $this->presenter->presentValue($subject),
-            $this->presenter->presentValue($arguments[0])
+            'Expected %s not to start with %s, but it does.',
+            $this->presenter->presentString($subject),
+            $this->presenter->presentString($arguments[0])
         ));
     }
 }
