+++ src/PhpSpec/Extensions/DefaultMatchers/Matcher/TraversableCountMatcher.php
@@ -11,10 +11,11 @@
  * file that was distributed with this source code.
  */
 
-namespace PhpSpec\Matcher;
+namespace PhpSpec\Extensions\DefaultMatchers\Matcher;
 
 use PhpSpec\Formatter\Presenter\Presenter;
 use PhpSpec\Exception\Example\FailureException;
+use PhpSpec\Matcher\Matcher;
 
 final class TraversableCountMatcher implements Matcher
 {
